Relocating the updated files
The game was on my PHP repository, but I decided to not use PHP but rather a API Word: https://random-word-api.herokuapp.com/home
The game is already working, but there are some glitches during its execution, in addition to not being intuitive and widely playable. I will fix them as soon as possible and update the files soon after.